% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ct_check.R
\name{Ct_check}
\alias{Ct_check}
\title{Ct_check}
\usage{
Ct_check(File_name)
}
\arguments{
\item{File_name}{File acquired from RT PCR}
}
\value{

}
\description{
Ct_check
}
\examples{
Ct_check("220621 inhibitory ACC calcium.xls")

}
